chore: consolidate evm trade exectution functions
