+++ src/lib/swapper/swappers/ThorchainSwapper/ThorchainSwapper2.ts
@@ -2,8 +2,6 @@ import type { AssetId } from '@shapeshiftoss/caip'
 import { CHAIN_NAMESPACE, fromChainId } from '@shapeshiftoss/caip'
 import type {
   CosmosSdkChainAdapter,
-  EvmChainAdapter,
-  EvmChainId,
   SignTx,
   UtxoChainAdapter,
   UtxoChainId,
@@ -12,7 +10,7 @@ import type { ThorchainSignTx } from '@shapeshiftoss/hdwallet-core'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
 import type { BuyAssetBySellIdInput, ExecuteTradeArgs, Swapper2 } from 'lib/swapper/api'
 import { assertUnreachable } from 'lib/utils'
-import { signAndBroadcast } from 'lib/utils/evm'
+import { executeEvmTrade } from 'lib/utils/evm'
 
 import { ThorchainSwapper } from './ThorchainSwapper'
 
@@ -24,12 +22,7 @@ export const thorchainSwapper: Swapper2 = {
 
     switch (chainNamespace) {
       case CHAIN_NAMESPACE.Evm: {
-        const evmChainAdapter = adapter as unknown as EvmChainAdapter
-        return signAndBroadcast({
-          adapter: evmChainAdapter,
-          txToSign: txToSign as SignTx<EvmChainId>,
-          wallet,
-        })
+        return executeEvmTrade({ txToSign, wallet, chainId })
       }
 
       case CHAIN_NAMESPACE.CosmosSdk: {
