+++ src/lib/swapper/swappers/ZrxSwapper/ZrxSwapper2.ts
@@ -1,9 +1,6 @@
 import type { AssetId } from '@shapeshiftoss/caip'
-import type { EvmChainAdapter } from '@shapeshiftoss/chain-adapters'
-import type { ETHSignTx } from '@shapeshiftoss/hdwallet-core'
-import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
-import type { BuyAssetBySellIdInput, ExecuteTradeArgs, Swapper2 } from 'lib/swapper/api'
-import { signAndBroadcast } from 'lib/utils/evm'
+import type { BuyAssetBySellIdInput, Swapper2 } from 'lib/swapper/api'
+import { executeEvmTrade } from 'lib/utils/evm'
 import { selectAssets } from 'state/slices/selectors'
 import { store } from 'state/store'
 
@@ -13,12 +10,7 @@ import type { ZrxSupportedChainId } from './types'
 import { ZRX_SUPPORTED_CHAINIDS, ZRX_UNSUPPORTED_ASSETS } from './utils/constants'
 
 export const zrxSwapper: Swapper2 = {
-  executeTrade: ({ txToSign, wallet, chainId }: ExecuteTradeArgs) => {
-    const adapterManager = getChainAdapterManager()
-    const adapter = adapterManager.get(chainId) as unknown as EvmChainAdapter
-
-    return signAndBroadcast({ adapter, wallet, txToSign: txToSign as ETHSignTx })
-  },
+  executeTrade: executeEvmTrade,
 
   filterAssetIdsBySellable: (assetIds: AssetId[]): Promise<AssetId[]> => {
     const assets = selectAssets(store.getState())
